Add new C files and headers for managing student database
