+++ src/lib/headers/db.h
@@ -1,0 +1,9 @@
+#ifndef DB_H
+#define DB_H
+
+#include <sqlite3.h>
+
+// initialize the database
+int init_db();
+
+#endif /* DB_H */
